Editing events

+++ Eventify/Controllers/EventsController.cs
@@ -62,11 +62,12 @@ namespace WebApplication2.Controllers
     public class EventsController : Controller
     {
         IEventService _manager;
-
         public EventsController(IEventService managerEvents)
         {
             _manager = managerEvents;
         }
+
+
 
         [HttpGet]
         public IActionResult Index()
@@ -79,7 +80,6 @@ namespace WebApplication2.Controllers
                 maxPrice: null,
                 startDate: null,
                 endDate: null,
-                isPrivate: null,
                 out int totalEvents
                 );
 
@@ -113,7 +113,6 @@ namespace WebApplication2.Controllers
                 maxPrice: eventBrowseViewModel.MaxPrice,
                 startDate: eventBrowseViewModel.StartDate,
                 endDate: eventBrowseViewModel.EndDate,
-                isPrivate: eventBrowseViewModel.IsPrivate,
                 out int totalEvents
                 );
 
@@ -136,29 +135,32 @@ namespace WebApplication2.Controllers
         }
 
 
+
         [HttpGet]
-        public IActionResult Add(int VenueId)
+        public IActionResult Add(int id)
         {
             if (User.Identity.IsAuthenticated && User.IsInRole("Organizer"))
             {
-                EventAddOrEditVM vm = new EventAddOrEditVM();
-                vm.VenueId = VenueId;
+                EventAddVM vm = new EventAddVM();
+                vm.VenueId = id;
                 return View(vm);
             }
             return RedirectToAction("Login", "Account");
         }
 
         [HttpPost]
-        public IActionResult Add(EventAddOrEditVM vm)
+        public IActionResult Add(EventAddVM vm)
         {
             var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
             if (!string.IsNullOrEmpty(userId) && User.IsInRole("Organizer"))
             {
+                List<string> uploadedFileNames = new List<string>();
                 foreach (var x in vm.FormFiles)
                 {
                     string p = UploadEventPhoto.UploadFile("images", x);
                     EventPhoto eventPhoto = new EventPhoto();
                     eventPhoto.PhotoUrl = $"/images/" + p;
+                    uploadedFileNames.Add(p);
                     vm.EventPhotos.Add(eventPhoto);
                 }
                 if (ModelState.IsValid)
@@ -175,17 +177,25 @@ namespace WebApplication2.Controllers
                         IsPrivate = vm.IsPrivate,
                         EventPhotos = vm.EventPhotos,
                         VenueId = vm.VenueId,
+                        Status = EventStatusEnum.Pending
                     };
 
                     _manager.Insert(ev, userId);
                     return RedirectToAction("Index");
                 }
-                return View(vm);
-
+                else 
+                {
+                    foreach (var file in uploadedFileNames)
+                    {
+                        UploadEventPhoto.RemoveFile("images", file);
+                    }
+                    vm.EventPhotos.Clear();
+                    return View(vm);
+                }
             }
-
             return RedirectToAction("Login", "Account");
         }
+
 
 
         public IActionResult Details(int id)
@@ -194,6 +204,7 @@ namespace WebApplication2.Controllers
             if (ev != null)
             {
                 EventDetailsVM vm = new EventDetailsVM();
+                vm.EventId = ev.EventId;
                 vm.EventTitle = ev.EventTitle;
                 vm.Description = ev.Description;
                 vm.StartDateTime = ev.StartDateTime;
@@ -204,6 +215,7 @@ namespace WebApplication2.Controllers
                 vm.Category = ev.Category;
                 vm.OrganizerName = ev.Organizer.UserName;
                 vm.OrganizerId = ev.OrganizerId;
+                vm.OwnerId = ev.Venue.OwnerId;
                 vm.Capacity = ev.Capacity;
                 vm.EventPhotos = ev.EventPhotos;
                 return View(vm);
@@ -211,6 +223,8 @@ namespace WebApplication2.Controllers
             }
             return NotFound();
         }
+
+
 
         [HttpGet]
         public IActionResult Edit(int id)
@@ -224,7 +238,7 @@ namespace WebApplication2.Controllers
                     Event ev = _manager.GetByIdWithIncludes(id);
                     if (ev != null)
                     {
-                        EventAddOrEditVM vm = new EventAddOrEditVM();
+                        EventEditVM vm = new EventEditVM();
                         vm.EventId = ev.EventId;
                         vm.EventTitle = ev.EventTitle;
                         vm.Category = ev.Category;
@@ -238,14 +252,16 @@ namespace WebApplication2.Controllers
                         {
                             vm.EventPhotos.Add(t);
                         }
+                        vm.OriginalPhotoCount = ev.EventPhotos.Count;
                         return View(vm);
                     }
                 }
             }
             return RedirectToAction("Login", "Account");
         }
+
         [HttpPost]
-        public IActionResult Edit(EventAddOrEditVM vm)
+        public IActionResult Edit(EventEditVM vm)
         {
             var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
             if (userId != null)
@@ -253,6 +269,13 @@ namespace WebApplication2.Controllers
                 int Id = int.Parse(userId);
                 if (User.IsInRole("Organizer"))
                 {
+                    foreach (var x in vm.FormFiles)
+                    {
+                        string p = UploadEventPhoto.UploadFile("images", x);
+                        EventPhoto eventPhoto = new EventPhoto();
+                        eventPhoto.PhotoUrl = $"/images/" + p;
+                        vm.EventPhotos.Add(eventPhoto);
+                    }
                     if (ModelState.IsValid)
                     {
                         var ev = _manager.GetByIdWithIncludes(vm.EventId);
@@ -268,6 +291,7 @@ namespace WebApplication2.Controllers
                         ev.TicketPrice = vm.TicketPrice;
                         ev.Features = vm.Features;
                         ev.IsPrivate = vm.IsPrivate;
+                        ev.Status = EventStatusEnum.Pending;
 
                         if (vm.DeletedPhotos != null && vm.DeletedPhotos.Any())
                         {
@@ -293,26 +317,41 @@ namespace WebApplication2.Controllers
                                     PhotoUrl = "/images/" + p
                                 });
                             }
-
                         }
-
-
                         _manager.Update(ev);
                         return RedirectToAction("Index");
                     }
                     return View(vm);
                 }
-                return RedirectToAction("Login", "Account");
             }
-            return NotFound();
+            return RedirectToAction("Login", "Account");
         }
-    
-    
+
+
+
+        public IActionResult Approval(int Id,EventStatusEnum decision)
+        {
+            Event ev = _manager.GetById(Id);
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (userId != null && int.Parse(userId) == ev.Venue.OwnerId)
+            {
+                ev.Status = decision;
+                _manager.Update(ev);
+            }
+            return RedirectToAction("Details", "Events", new { id = Id });
+        }
+
+
+
         public IActionResult Delete(int Id)
         {
-            _manager.Delete(Id);
-            return View("Index");
+            Event ev = _manager.GetByIdWithIncludes(Id);
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (userId != null && int.Parse(userId) == ev.OrganizerId)
+            {
+                _manager.Delete(Id);
+            }
+            return RedirectToAction("Index", "Events");
         }
-    
     }
 }
